LPS-179567 Extract CPInstanceTierPriceEntriesScreenNavigationCategory methods to a CPInstanceTierPriceEntriesScreenNavigationEntry component and extends CPInstanceTierPriceEntriesScreenNavigationCategory in it
